+++ upgrade/upgrade-2.3.25.php
@@ -1,0 +1,10 @@
+<?php
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+function upgrade_module_2_3_25($module)
+{
+    return true;
+}